+++ DeepAzureServer/Models/Entities/Message.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Content { get; set; }
         public int ConversationId { get; set; }
-        public int SenderId { get; set; }
+        public long SenderId { get; set; }
 
         public Conversation? Conversation { get; set; }
         public User? Sender { get; set; }
